SQL lite has been integrated with the project. Some tables also contains some data. Need to work on student table.

+++ src/model/entities/Team.java
@@ -272,7 +272,7 @@ public class Team implements Serializable{
 											.mapToDouble(Double::valueOf)
 											.sum();
 		
-		this.totalSkillShortage = (double) totalSkillShortage;
+		this.totalSkillShortage = totalSkillShortage;
 	}
